separate books and chapters layers + create ChapterContent model

+++ src/main/java/com/mythiqa/mythiqabackend/dto/response/NewChapterNumDTO.java
@@ -1,3 +1,3 @@
 package com.mythiqa.mythiqabackend.dto.response;
 
-public record NewChapterNumDTO (int chapterNumber) {}
+public record NewChapterNumDTO (long chapterNumber) {}
